--- .github/pull_request_template.md
@@ -1,17 +1,0 @@
-## 概要
-
-このセクションでは、この PR の目的と概要を簡潔に説明してください。
-
-## 変更点
-
-このセクションでは、具体的な変更点や修正箇所を箇条書きでリストアップしてください。
-
-- 変更点 1
-- 変更点 2
-- 変更点 3
-
-## 関連 Issue
-
-このセクションでは、この PR が関連する Issue やタスクをリンクしてください。以下のように記述します。
-
-- 関連 Issue: #123
