--- .vscode/extensions.json
@@ -1,5 +1,0 @@
-{
-  "recommendations": [
-    "esbenp.prettier-vscode" // prettier
-  ]
-}
