+++ index.html
@@ -15,10 +15,8 @@
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link href="https://fonts.googleapis.com/css2?family=Shippori+Mincho:wght@400;500;600;700;800&display=swap" rel="stylesheet" />
-    <link rel="stylesheet" href="styles/reset.css" />
-    <link rel="stylesheet" href="styles/style.css" />
-    <script src="./scripts/main.js" defer></script>
-    <script src="./scripts/explanation.js" defer></script>
+    <link rel="stylesheet" href="src/styles/reset.css" />
+    <link rel="stylesheet" href="src/styles/style.css" />
     <title>team-c</title>
   </head>
   <body id="js-layout" class="layout">
@@ -32,5 +30,7 @@
     <div id="js-explanation"></div>
     <div id="js-game"></div>
     <p id="js-score" class="score"></p>
+    <script src="src/scripts/main.js" type="module"></script>
+    <script src="src/scripts/explanation.js" type="module"></script>
   </body>
 </html>
