+++ vite.config.js
@@ -1,0 +1,5 @@
+import { defineConfig } from "vite";
+
+export default defineConfig({
+  base: "team-c",
+});
